Move Feeds related type matchers to service/request/shape
